+++ libyaul/kernel/dbgio/devices/vdp2.c
@@ -46,13 +46,13 @@ typedef struct {
 static void _init(const dbgio_vdp2_t *);
 static void _flush(void);
 
-static inline void __attribute__ ((always_inline)) _pnd_clear(uint32_t, uint32_t);
-static inline void __attribute__ ((always_inline)) _pnd_write(uint32_t, uint32_t, uint16_t);
+static inline void __attribute__ ((always_inline)) _pnd_clear(int16_t, int16_t);
+static inline void __attribute__ ((always_inline)) _pnd_write(int16_t, int16_t, uint16_t);
 
 static void _buffer_clear(void);
-static void _buffer_area_clear(int32_t, int32_t, int32_t, int32_t);
-static void _buffer_line_clear(int32_t, int32_t, int32_t);
-static void _buffer_write(int32_t, int32_t, uint8_t);
+static void _buffer_area_clear(int16_t, int16_t, int16_t, int16_t);
+static void _buffer_line_clear(int16_t, int16_t, int16_t);
+static void _buffer_write(int16_t, int16_t, uint8_t);
 
 static void _dma_font_handler(void *);
 static void _dma_handler(void *);
@@ -284,15 +284,15 @@ _flush(void)
 }
 
 static inline void __attribute__ ((always_inline))
-_pnd_clear(uint32_t col, uint32_t row)
+_pnd_clear(int16_t col, int16_t row)
 {
         _pnd_write(col, row, _dev_state->pnd_clear);
 }
 
 static inline void __attribute__ ((always_inline))
-_pnd_write(uint32_t col, uint32_t row, uint16_t value)
+_pnd_write(int16_t col, int16_t row, uint16_t value)
 {
-        uint32_t offset;
+        int16_t offset;
         offset = col + (row * _dev_state->page_width);
 
         _dev_state->page_pnd[offset] = value;
@@ -305,14 +305,14 @@ _buffer_clear(void)
 }
 
 static void
-_buffer_area_clear(int32_t col_start, int32_t col_end, int32_t row_start,
-    int32_t row_end)
+_buffer_area_clear(int16_t col_start, int16_t col_end, int16_t row_start,
+    int16_t row_end)
 {
         _dev_state->state |= STATE_BUFFER_DIRTY;
 
-        int32_t row;
+        int16_t row;
         for (row = row_start; row < row_end; row++) {
-                int32_t col;
+                int16_t col;
                 for (col = col_start; col < col_end; col++) {
                         _pnd_clear(col, row);
                 }
@@ -320,18 +320,18 @@ _buffer_area_clear(int32_t col_start, int32_t col_end, int32_t row_start,
 }
 
 static void
-_buffer_line_clear(int32_t col_start, int32_t col_end, int32_t row)
+_buffer_line_clear(int16_t col_start, int16_t col_end, int16_t row)
 {
         _dev_state->state |= STATE_BUFFER_DIRTY;
 
-        int32_t col;
+        int16_t col;
         for (col = col_start; col < col_end; col++) {
                 _pnd_clear(col, row);
         }
 }
 
 static void
-_buffer_write(int32_t col, int32_t row, uint8_t ch)
+_buffer_write(int16_t col, int16_t row, uint8_t ch)
 {
         _dev_state->state |= STATE_BUFFER_DIRTY;
